--- src/containers/Measurement.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Measurement = () => {
-  return <div></div>;
-};
-
-export default Measurement;
